add docs cross_correlation rename to cross_correlation
