+++ pastas_plugins/cross_correlation/__init__.py
@@ -1,0 +1,4 @@
+# ruff: noqa: F401
+from pastas_plugins.cross_correlation.crosscorr import ccf, fit_response, prewhiten
+from pastas_plugins.cross_correlation.plots import plot_ccf_overview, plot_corr
+from pastas_plugins.cross_correlation.version import __version__
