--- pastas_plugins/crosscorrelation/__init__.py
@@ -1,4 +1,0 @@
-# ruff: noqa: F401
-from pastas_plugins.crosscorrelation.crosscorr import ccf, fit_response, prewhiten
-from pastas_plugins.crosscorrelation.plots import ccf_overview_plot, plot_corr
-from pastas_plugins.crosscorrelation.version import __version__
